requirements.txt
Flask
Flask-Migrate
Flask-SQLAlchemy
mysqlclient
SQLAlchemy
python-dotenv
requests
python-json-logger
